simulate_devices/requirements.txt
paho-mqtt>=1.6.0
python-dotenv>=1.0.0
